% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{Acc}
\alias{Acc}
\title{Computes the proportion of matching terms in two vectors of the same length.
Used to compute the accuracy for prediction on test set.}
\usage{
Acc(Vec1, Vec2)
}
\arguments{
\item{Vec1}{A vector of labels}

\item{Vec2}{Another vector of labels}
}
\value{
Percentage of identical labels (accuracy)
}
\description{
Computes the proportion of matching terms in two vectors of the same length.
Used to compute the accuracy for prediction on test set.
}
\examples{
Vec1 <- c(1,1,2,3,1)
Vec2 <- c(1,2,2,3,1)
Acc(Vec1,Vec2)
}
