% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.R
\name{BEST}
\alias{BEST}
\title{Main function of the package.
It produces Classification Trees with Branch-Exclusive variables.}
\usage{
BEST(Data, Size, VA)
}
\arguments{
\item{Data}{A data set (Data Frame): Can take on both numerical and categorical predictors. Last column of the data set must be the Repsonse Variable (Categorical Variables only)}

\item{Size}{Minimal Number of Observation within a leaf needed for partitionning}

\item{VA}{Variable Availability structure}
}
\value{
A BEST object with is a list containing the resulting tree, row numbers for each regions and the split points
}
\description{
Main function of the package.
It produces Classification Trees with Branch-Exclusive variables.
}
\examples{
n <- 1000
Data <- BESTree::Data[1:n,]
d <- ncol(Data)-1
VA <- ForgeVA(d,1,0,0,0)
Size <- 50
Fit <- BESTree::BEST(Data,Size,VA)
}
