% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Data generated according to decision tree for simulation purposes}
\format{A data frame with 10000 rows and 5 variables:
\describe{
  \item{X_1}{Binary predictor}
  \item{X_2}{Binary predictor}
  \item{X_3}{Continuous predictor between 0 and 1}
  \item{X_4}{Continuous predictor between 0 and 1}
  \item{Y}{The response variable}
  ...
}}
\usage{
Data
}
\description{
Data generated according to decision tree for simulation purposes
}
\keyword{datasets}
