% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\docType{data}
\name{Fit}
\alias{Fit}
\title{Data generated according to decision tree for simulation purposes}
\format{A typical list produced by the BEST function:
\describe{
  \item{1}{Tree structure indicating spliting variables, impurity of the region and split variable}
  \item{2}{List of splitting values}
  \item{3}{Observaton numbers in the respective regions}
  ...
}}
\usage{
Fit
}
\description{
Data generated according to decision tree for simulation purposes
}
\keyword{datasets}
