% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestforest.R
\name{FPredict}
\alias{FPredict}
\title{Emits prediction from a forest of BEST's}
\usage{
FPredict(M, LFit)
}
\arguments{
\item{M}{A matrix of new observations where one row is one observation}

\item{LFit}{A list of BEST Objects (Usually produced by RBEST or BESTForest)}
}
\value{
A vector of predictions
}
\description{
Emits prediction from a forest of BEST's
}
\examples{
n <- 500
Data <- BESTree::Data[1:n,]
d <- ncol(Data)-1
NewPoints <- BESTree::Data[(n+1):(n+11),1:d]
VA <- ForgeVA(d,1,0,0,0)
Size <- 50
NoT <- 10
Fit <- BESTree::BaggedBEST(Data,VA,NoT,Size)
Predictions <- BESTree::FPredict(NewPoints,Fit)
}
