% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.R
\name{Predict}
\alias{Predict}
\title{Classify a new observation point}
\usage{
Predict(Point, Fit)
}
\arguments{
\item{Point}{A new observation}

\item{Fit}{A BEST object}
}
\value{
The predicted class
}
\description{
Classify a new observation point
}
\examples{
n <- 500
Data <- BESTree::Data[1:n,]
NewPoint <- BESTree::Data[n+1,]
d <- ncol(Data)-1
VA <- ForgeVA(d,1,0,0,0)
Size <- 50
Fit <- BESTree::BEST(Data,Size,VA)
BESTree::Predict(NewPoint[1:d],Fit)
}
