% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.R
\name{TreePruning}
\alias{TreePruning}
\title{Uses a Validation Set to select the best trees within the list of pruned trees.}
\usage{
TreePruning(Fit, VSet)
}
\arguments{
\item{Fit}{A BEST object}

\item{VSet}{A Validation Set (Can also be used in CV loop)}
}
\value{
The shallower trees among trees wiht Highest accuracy. This replaces the first element in the BEST object list.
}
\description{
Uses a Validation Set to select the best trees within the list of pruned trees.
}
\examples{
nv <- 50
ValData <- BESTree::Data[(1000+1):nv,]
Fit <- BESTree::Fit
Fit[[1]] <- BESTree::TreePruning(Fit,ValData)
}
